backend/requirements.txt
torch
transformers
google-generativeai
python-dotenv
